Test is now passing
